+++ test/test_proj.py
@@ -117,7 +117,7 @@ class TypeError_Transform_Issue8_Test(unittest.TestCase):
     # https://github.com/jswhit/pyproj/issues/8
 
     def setUp(self):
-        with pytest.warns(DeprecationWarning):
+        with pytest.warns(FutureWarning):
             self.p = Proj(init="epsg:4269")
 
     def test_tranform_none_1st_parmeter(self):
@@ -367,12 +367,6 @@ class Geod_NaN_Issue112_Test(unittest.TestCase):
         self.assertTrue(lon2 != lon2)
         self.assertTrue(lat2 == lat2)
         self.assertTrue(azi2 == azi2)
-
-
-def test_proj_version():
-    awips221 = Proj(proj="lcc", R=6371200, lat_1=50, lat_2=50, lon_0=-107)
-    with pytest.warns(DeprecationWarning, match="'Proj.proj_version' is deprecated"):
-        assert type(awips221.proj_version) is int
 
 
 def test_proj_equals():